+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - Add Sandbox Messages API
 - Add Sending Domains API
+- Add Account Accesses API
 
 ## [2.6.0] - 2026-01-27
 - Add Inboxes API